+++ release_notes.md
@@ -129,5 +129,6 @@ First release into griffon artifacts portal.
 
 ### 0.7
 1. Add `findAllModelFetch[Graph]()`, `findModelByDslFetch[Graph]()`, and `findModelByXXXFetch[Graph]()`.
+1. Domain classes now can be in any packages.
 1. Fixes `@Transaction` to restrictive: allow transformation on getter and setter.
 1. Finders by default will remove duplicate entities.  Use `allowDuplicate: true` config or global `griffon.simplejpa.finders.alwaysAllowDuplicate` config to allow duplicate entities in the result.